requirements.txt
beautifulsoup4==4.6.0
certifi==2017.11.5
chardet==3.0.4
idna==2.6
lyricwikia==0.1.4
numpy==1.13.3
pandas==0.21.0
python-dateutil==2.6.1
python-decouple==3.1
pytz==2017.3
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
sklearn==0.0
spotipy==2.4.4
urllib3==1.22
